✅ done: spawn/despawn 확인 완료

+++ Server/ServerCore/SendBuffer.h
@@ -19,9 +19,9 @@ public:
 
 					/* SendBuffer 정보 관련 */
 	BYTE*			Buffer()							{ return _buffer.data(); }
-	int32			Len()								{ return _buffer.size(); }
+	int32			Len()								{ return (int32)_buffer.size(); }
 	int32			WriteSize()							{ return _writePos; }
-	int32			FreeSize()							{ return _buffer.size() - _writePos; }
+	int32			FreeSize()							{ return (int32)_buffer.size() - _writePos; }
 	BYTE*			WritePos()							{ return &_buffer[_writePos]; }
 
 					/* RecvBuffer 데이터 조작 관련 */
